Last part of project, refreshing laravel knowledge completed, maybe I will make better layout in the future

+++ routes/web.php
@@ -24,24 +24,26 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', [RecipesController::class, 'index']);
 
-Route::get('/recipe/{recipe}/edit', [RecipesController::class, 'EditRecipe']);
+Route::get('/recipe/{recipe}/edit', [RecipesController::class, 'EditRecipe'])->middleware('auth');
 
-Route::put('/recipe/{recipe}', [RecipesController::class, 'UpdateRecipe']);
+Route::put('/recipe/{recipe}', [RecipesController::class, 'UpdateRecipe'])->middleware('auth');
 
-Route::delete('/recipe/{recipe}', [RecipesController::class, 'DestroyRecipe']);
+Route::delete('/recipe/{recipe}', [RecipesController::class, 'DestroyRecipe'])->middleware('auth');
 
 Route::get('/recipe/{recipe}', [RecipesController::class, 'SingleRecipe']);
 
-Route::post('/create', [RecipesController::class, 'Save']);
+Route::post('/create', [RecipesController::class, 'Save'])->middleware('auth');
 
-Route::get('/create', [RecipesController::class, 'Create']);
+Route::get('/create', [RecipesController::class, 'Create'])->middleware('auth');
+
+Route::get('/profile', [UserController::class, 'userProfile'])->middleware('auth');
 
 Route::post('/register', [UserController::class, 'CreateUser']);
 
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 Route::post('/login', [UserController::class, 'authenticate']);
 
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
-Route::get('/register', [UserController::class, 'registerUser']);
+Route::get('/register', [UserController::class, 'registerUser'])->middleware('guest');
